misc: reorganize package
