+++ src/main/java/org/aya/intellij/actions/FindUsages.java
@@ -3,8 +3,8 @@ package org.aya.intellij.actions;
 import com.intellij.lang.cacheBuilder.WordsScanner;
 import com.intellij.lang.findUsages.FindUsagesProvider;
 import com.intellij.psi.PsiElement;
-import org.aya.intellij.parser.AyaParserDefinition;
-import org.aya.intellij.parser.AyaWordsScanner;
+import org.aya.intellij.language.AyaParserDefinition;
+import org.aya.intellij.language.AyaWordsScanner;
 import org.aya.intellij.psi.AyaPsiNamedElement;
 import org.aya.intellij.psi.concrete.*;
 import org.jetbrains.annotations.Nls;
