+++ src/main/java/org/aya/intellij/actions/lsp/JB.java
@@ -1,4 +1,4 @@
-package org.aya.intellij.lsp;
+package org.aya.intellij.actions.lsp;
 
 import com.intellij.diff.util.LineCol;
 import com.intellij.openapi.project.Project;
