+++ src/main/java/org/aya/intellij/inspection/CatchAll.java
@@ -5,8 +5,8 @@ import com.intellij.codeInspection.ProblemsHolder;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.PsiFile;
 import kala.collection.SeqView;
-import org.aya.intellij.lsp.AyaLsp;
-import org.aya.intellij.lsp.JB;
+import org.aya.intellij.actions.lsp.AyaLsp;
+import org.aya.intellij.actions.lsp.JB;
 import org.aya.intellij.psi.concrete.AyaPsiVisitor;
 import org.aya.intellij.service.DistillerService;
 import org.aya.util.reporter.Problem;
