+++ src/main/java/org/aya/intellij/inspection/error/ErrorInspection.java
@@ -8,10 +8,10 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import kala.collection.SeqView;
 import org.aya.intellij.AyaBundle;
+import org.aya.intellij.actions.lsp.AyaLsp;
+import org.aya.intellij.actions.lsp.JB;
 import org.aya.intellij.inspection.CatchAll;
 import org.aya.intellij.inspection.Jobs;
-import org.aya.intellij.lsp.AyaLsp;
-import org.aya.intellij.lsp.JB;
 import org.aya.intellij.psi.AyaPsiFile;
 import org.aya.intellij.service.DistillerService;
 import org.aya.util.reporter.Problem;
