+++ src/main/java/org/aya/intellij/inspection/warning/DominationInspection.kt
@@ -3,7 +3,7 @@ package org.aya.intellij.inspection.warning
 import com.intellij.codeInspection.*
 import com.intellij.openapi.project.Project
 import org.aya.intellij.AyaBundle
-import org.aya.intellij.lsp.AyaLsp
+import org.aya.intellij.actions.lsp.AyaLsp
 import org.aya.intellij.psi.AyaPsiElement
 import org.aya.intellij.psi.concrete.AyaPsiBareClause
 import org.aya.intellij.psi.concrete.AyaPsiBarredClause
