+++ src/main/java/org/aya/intellij/language/AyaNamesValidator.java
@@ -1,4 +1,4 @@
-package org.aya.intellij.parser;
+package org.aya.intellij.language;
 
 import com.intellij.lang.refactoring.NamesValidator;
 import com.intellij.openapi.project.Project;
