+++ src/main/java/org/aya/intellij/ui/settings/AyaColorSettingsPage.java
@@ -6,8 +6,8 @@ import com.intellij.openapi.options.colors.AttributesDescriptor;
 import com.intellij.openapi.options.colors.ColorDescriptor;
 import com.intellij.openapi.options.colors.ColorSettingsPage;
 import com.intellij.openapi.util.NlsContexts;
-import org.aya.intellij.AyaIcons;
 import org.aya.intellij.actions.SyntaxHighlight;
+import org.aya.intellij.ui.AyaIcons;
 import org.aya.parser.AyaLanguage;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
