+++ src/main/java/org/aya/intellij/ui/toolwindow/AyaToolWindow.java
@@ -3,8 +3,8 @@ package org.aya.intellij.ui.toolwindow;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowFactory;
-import org.aya.intellij.AyaIcons;
 import org.aya.intellij.service.ProblemService;
+import org.aya.intellij.ui.AyaIcons;
 import org.jetbrains.annotations.NotNull;
 
 public class AyaToolWindow implements ToolWindowFactory {
