+++ src/main/java/org/aya/intellij/ui/toolwindow/GoalsView.java
@@ -18,12 +18,12 @@ import kotlin.Unit;
 import org.aya.core.term.ErrorTerm;
 import org.aya.core.term.Term;
 import org.aya.intellij.AyaBundle;
-import org.aya.intellij.AyaIcons;
-import org.aya.intellij.lsp.AyaLsp;
-import org.aya.intellij.lsp.JB;
+import org.aya.intellij.actions.lsp.AyaLsp;
+import org.aya.intellij.actions.lsp.JB;
 import org.aya.intellij.psi.AyaPsiFile;
 import org.aya.intellij.psi.concrete.AyaPsiHoleExpr;
 import org.aya.intellij.service.ProblemService;
+import org.aya.intellij.ui.AyaIcons;
 import org.aya.intellij.ui.AyaTreeView;
 import org.aya.pretty.doc.Doc;
 import org.aya.tyck.error.Goal;
